Fix warnings.

+++ lcfit_src/lcfit.h
@@ -212,7 +212,7 @@ void lcfit_bsm_rescale(const double t, const double l, bsm_t* m);
  * \return An #lcfit_status code, zero for success, non-zero otherwise.
  */
 int lcfit_fit_bsm_weight(const size_t n, const double* t, const double* l,
-                         const double* w, bsm_t* m, int max_iter);
+                         const double* w, bsm_t* m, size_t max_iter);
 
 /** Fit a given model to empirical likelihood data without weighting.
  *
